application/translation-service/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
openai-whisper
ffmpeg-python
deep-translator